+++ src/App.js
@@ -9,8 +9,7 @@ import NavbarMenu from './navbar/Navbar.js';
 // Components
 import About from './views/About/About.js';
 import Gallary from './views/Image_Gallary/ImageList.js'
-import Footer from './views/Footer/Footer.js';
-
+import ContactForm from './views/Contact/Contact.js';
  
 class App extends React.Component {
   constructor(props) {
@@ -49,12 +48,9 @@ class App extends React.Component {
 
       <Container>
         < About />
-      
         <Gallary images={this.state.images} />
-        
+        < ContactForm / >
       </Container>
-
-      <Footer />
     </div>
   )
   }
