--- src/views/Footer/Footer.css
@@ -1,28 +1,0 @@
-.footer_component {
-    margin-top: 50px;
-    height: 500px;
-    background-color: rgb(195, 185, 185);
-}
-
-.upperFooter {
-    height: 85%; 
-    background-color: #f1f1f1;
-}
-
-.bottomFooter {
-    height: 15%; 
-    background-color: #717171;
-}
-
-.contact_header {
-    text-align: center;
-    padding-top: 50px;
-}
-
-.contact_form {
-    border: 2px solid red;
-    width: 60%;
-    margin: 0 auto;
-    background-color: rgb(192, 77, 77);
-    height: 100%;
-}
