+++ src/views/Footer/Footer.js
@@ -1,15 +1,10 @@
 import React from 'react';
 import './Footer.css';
 
-import ContactForm from './Contact/Contact.js'
-
 const Footer = (props) => (
     <div className='footer_component'>
-        <div className="upperFooter">
-            <ContactForm />
-        </div>
-        <div className="bottomFooter">
-
+        <div className="footer">
+            <p> Footer </p>
         </div>
     </div>
 )
