+++ src/views/Image_Gallary/ImageList.css
@@ -1,9 +1,7 @@
 .gallaryImages {
-    border: 1px solid rgb(0, 0, 0);
-    margin: 0 auto;
-    margin-top: 50px;
     text-align: center;
-    
+    padding-top: 50px;
+    padding-bottom: 50px;
 }
 
 .gallaryImages img {
